refactor: rename visitors per centage summary pie chart to bar chart
