+++ src/lib/components/chart/config.ts
@@ -1,14 +1,14 @@
 import type { VisitorsSummary } from '@data';
-import type { ChartConfiguration, ChartData, ChartItem, ChartType, LabelItem, TooltipItem } from 'chart.js';
+import type { ChartConfiguration, ChartData, ChartType, TooltipItem } from 'chart.js';
 
-export function visitorsPercentageSummaryPieChartConfig(
+export function visitorsPercentageSummaryBarChartConfig(
 	summary: VisitorsSummary
 ): ChartConfiguration {
 	const data = <ChartData>{
 		labels: ['Mobile', 'Non-Mobile'],
 		datasets: [
 			{
-				data: [summary.mobileVisitorsPercentage, summary.nonMobileVisitorsPercentage],
+				data: [summary.allMobileVisitorsPercentage, summary.allNonMobileVisitorsPercentage],
 				backgroundColor: ['rgb(255, 99, 132)', 'rgb(54, 162, 235)'],
 				hoverOffset: 41,
 				borderRadius: 5
